+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Default headers: connection, accept, user-agent
 - Can provide options on each verb method
 - Support HTTP requests without curl, using file_get_contents()
+- Use content-type application/x-www-form-urlencoded when sending body content
 
 ## 0.0.1 : 2023-12-02
 - Basic working version